Add truncate to string/slice utilities
